Address linting concerns
- Broke tests out into separate files.
- Reduced text complexity.
- Reduced code duplication.
